rewrite_function_untested
